containers/langchain-worker/requirements.txt
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.3
httpx==0.28.1
redis==5.2.1
psycopg2-binary==2.9.10
pgvector==0.3.6
pydantic==2.10.4
pyyaml==6.0.2
